DabaseConncetor/requirements.txt
Python 3.12+
psycopg2-binary
python-dotenv
